requirements.txt
django
psycopg2-binary
pillow
django-jazzmin
